test: 테스트 코드 이동
서비스 테스트에 있던, 리포지토리에 대한 테스트코드를 옮겼고 단위 테스트에 더 알맞게 변경
